Extract celo functions into celo_state_transition.go
